VirusLens/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
pandas==2.2.3
sqlalchemy==2.0.34
tldextract==5.1.2
